Move auto download logic out of use case
